# ios/BeGit/BeGit/openapi.yaml
components:
  schemas:
...
    handler.ConfirmPostRequest:
      properties:
        body:
          type: string
      type: object
...
    handler.CreatePostRequest:
      properties:
        body:
          type: string
        commit_sha:
          type: string
        content_source:
          default: github
          enum:
          - github
          - manual
          type: string
        github_login:
          type: string
        notification_id:
          type: integer
        post_type:
          default: commit
          enum:
          - commit
          - pull_request
          - memo
          type: string
        pull_request_number:
          type: integer
        repo_full_name:
          type: string
      type: object
...
    handler.ErrorResponse:
      properties:
        error:
          type: string
      type: object
...
    handler.PhotoJSON:
      properties:
        id:
          type: integer
        photo_type:
          type: string
        url:
          type: string
      type: object
    handler.PostFeedJSON:
      properties:
        additions:
          type: integer
        avatar_url:
          type: string
        body:
          type: string
        commit_count:
          type: integer
        created_at:
          type: string
        deletions:
          type: integer
        id:
          type: integer
        latest_commit_message:
          type: string
        login:
          type: string
        photos:
          items:
            $ref: '#/components/schemas/handler.PhotoJSON'
          type: array
          uniqueItems: false
        post_type:
          type: string
        repo_full_name:
          type: string
        status:
          type: string
        user_id:
          type: integer
      type: object
    handler.PostJSON:
      properties:
        additions:
          type: integer
        body:
          type: string
        commit_count:
          type: integer
        created_at:
          type: string
        deletions:
          type: integer
        id:
          type: integer
        latest_commit_message:
          type: string
        post_type:
          type: string
        repo_full_name:
          type: string
        status:
          type: string
        user_id:
          type: integer
      type: object
    handler.PostListResponse:
      properties:
        posts:
          items:
            $ref: '#/components/schemas/handler.PostFeedJSON'
          type: array
          uniqueItems: false
      type: object
...
  securitySchemes:
    BearerAuth:
      description: '`Authorization: Bearer <token>` 形式でアクセストークンを付与する。'
      in: header
      name: Authorization
      type: apiKey
info:
  description: BeGit バックエンド API。GitHub と連携したリポジトリ単位のグループ・通知・投稿機能を提供する。
  title: BeGit API
  version: "1.0"
openapi: 3.1.0
paths:
...
  /groups/{id}/posts:
    get:
      parameters:
      - description: グループ ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.PostListResponse'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: Unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: フィード取得
      tags:
      - posts
    post:
      parameters:
      - description: グループ ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/handler.CreatePostRequest'
                description: 投稿内容
                summary: request
        description: 投稿内容
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.PostJSON'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: Internal Server Error
        "502":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: Bad Gateway
      security:
      - BearerAuth: []
      summary: 投稿作成
      tags:
      - posts
  /groups/{id}/posts/{postId}:
    delete:
      parameters:
      - description: グループ ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      - description: 投稿 ID
        in: path
        name: postId
        required: true
        schema:
          type: integer
      responses:
        "204":
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: Internal Server Error
        "502":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: Bad Gateway
      security:
      - BearerAuth: []
      summary: 投稿削除
      tags:
      - posts
...
  /groups/{id}/posts/{postId}/confirm:
    post:
      description: 下書きを確定し is_draft を解除する。べき等。
      parameters:
      - description: グループ ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      - description: 投稿 ID
        in: path
        name: postId
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/handler.ConfirmPostRequest'
                description: 確定内容（本文の上書き等）
                summary: request
        description: 確定内容（本文の上書き等）
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.PostJSON'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: Not Found
      security:
      - BearerAuth: []
      summary: 下書き確定
      tags:
      - posts
  /groups/{id}/posts/{postId}/draft:
    get:
      description: ② Nice Work! 通知の draft_post_id に対応する下書きを取得する
      parameters:
      - description: グループ ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      - description: 投稿 ID
        in: path
        name: postId
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.PostJSON'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: Not Found
      security:
      - BearerAuth: []
      summary: 下書き取得
      tags:
      - posts
...
servers:
- url: /